Add SDOUtils into build and fix build error
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/cpp/sca@556107 13f79535-47bb-0310-9956-ffa450edef68

+++ runtime/core/src/tuscany/sca/util/SDOUtils.cpp
@@ -190,7 +190,7 @@ namespace tuscany
                                 case Type::DateType:
                                     {
                                         const SDODate date = sequence->getDateValue(i);
-                                        result = visitor.visit(dataObject, seqProperty, (const time_t)date.getTime());
+                                        result = visitor.visit(dataObject, seqProperty, (long)(const time_t)date.getTime());
                                     }
                                     break;
                                 case Type::BigDecimalType: 
@@ -307,7 +307,7 @@ namespace tuscany
                                         case Type::DateType:
                                             {
                                                 const SDODate date = doList.getDate(li);
-                                                result = visitor.visit(dataObject, pl[i], (const time_t)date.getTime());
+                                                result = visitor.visit(dataObject, pl[i], (long)(const time_t)date.getTime());
                                             }
                                             break;
                                         case Type::BigDecimalType: 
@@ -389,7 +389,7 @@ namespace tuscany
                                     case Type::DateType:
                                         {
                                             const SDODate date = dataObject->getDate(pl[i]);
-                                            result = visitor.visit(dataObject, pl[i], (const time_t)date.getTime());
+                                            result = visitor.visit(dataObject, pl[i], (long)(const time_t)date.getTime());
                                         }
                                         break;
                                     case Type::BigDecimalType:
